feat: Login verification

+++ blueprint.py
@@ -8,6 +8,13 @@ bp = Blueprint('main', __name__, url_prefix='/api')
 conn = pymongo.MongoClient("localhost", 27017)
 db = conn.thoth_db
 
+def verify(db):
+    col = db.user
+    encoded = request.cookies.get('accessToken')
+    decoded = jwt.decode(encoded, 'JEfWefI0E1qlnIz06qmob7cZp5IzH/i7KwOI2xqWfhE=', algorithms=["HS256"])
+    result = col.find_one({'_id': ObjectId(decoded['uid'])})
+    return result
+
 @bp.route('/notes', methods=['GET'])
 def get():
     col = db.note
@@ -16,15 +23,22 @@ def get():
 
 @bp.route('/notes', methods=['POST'])
 def post():
+    col = db.user
+    encoded = request.cookies.get('accessToken')
+    decoded = jwt.decode(encoded, 'JEfWefI0E1qlnIz06qmob7cZp5IzH/i7KwOI2xqWfhE=', algorithms=["HS256"])
+    result = col.find_one({'_id': ObjectId(decoded['uid'])})
+    if result:
     col = db.note
-    note = {}
-    note['uid'] = request.form['uid']
-    note['title'] = request.form['title']
-    note['code'] = request.form['code']
-    note['tag'] = request.form['tag']
-    note['ref'] = request.form['ref']
-    col.insert_one(note)
-    return dumps({'error':None})
+        note = {}
+        note['uid'] = decoded['uid']
+        note['title'] = request.form['title']
+        note['code'] = request.form['code']
+        note['tag'] = request.form['tag']
+        note['ref'] = request.form['ref']
+        col.insert_one(note)
+        return dumps({'error':None})
+    else:
+        return dumps({'error':'LoginNotVerified'})
 
 @bp.route('/signin', methods=['POST'])
 def signin():
@@ -58,33 +72,45 @@ def signup():
 
 @bp.route('/note', methods=['PUT'])
 def update():
-    col = db.note
-    note = {}
-    nid = request.form['nid']
-    note['title'] = request.form['title']
-    note['code'] = request.form['code']
-    note['tag'] = request.form['tag']
-    note['ref'] = request.form['ref']
-    result = col.update_one({'_id': ObjectId(nid)}, {"$set":note})
-    return dumps({'error':None})
+    if verify(db):
+        col = db.note
+        note = {}
+        nid = request.form['nid']
+        note['title'] = request.form['title']
+        note['code'] = request.form['code']
+        note['tag'] = request.form['tag']
+        note['ref'] = request.form['ref']
+        result = col.update_one({'_id': ObjectId(nid)}, {"$set":note})
+        return dumps({'error':None})
+    else:
+        return dumps({'error':'LoginNotVerified'})
 
 @bp.route('/note/<nid>', methods=['DELETE'])
 def delete(nid):
-    col = db.note
-    col.delete_one({'_id': ObjectId(nid)})
-    return dumps({'error':None})
+    if verify(db):
+        col = db.note
+        col.delete_one({'_id': ObjectId(nid)})
+        return dumps({'error':None})
+    else:
+        return dumps({'error':'LoginNotVerified'})
 
 @bp.route('/myboard', methods=['POST'])
 def scrab():
-    col = db.note
-    nid = request.form['nid']
-    result = col.find_one({'_id': ObjectId(nid)})
-    col = db.scrab
-    col.insert(result)
-    return dumps({'error':None})
+    if verify(db):
+        col = db.note
+        nid = request.form['nid']
+        result = col.find_one({'_id': ObjectId(nid)})
+        col = db.scrab
+        col.insert(result)
+        return dumps({'error':None})
+    else:
+        return dumps({'error':'LoginNotVerified'})
 
 @bp.route('/myboard/<nid>', methods=['DELETE'])
 def unscrab(nid):
-    col = db.scrab
-    col.delete_one({'_id': ObjectId(nid)})
-    return dumps({'error':None})
+    if verify(db):
+        col = db.scrab
+        col.delete_one({'_id': ObjectId(nid)})
+        return dumps({'error':None})
+    else:
+        return dumps({'error':'LoginNotVerified'})
